% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filters.R
\name{mod_filters_ui}
\alias{mod_filters_ui}
\alias{mod_filters_server}
\title{mod_filters_ui and mod_filters_server}
\usage{
mod_filters_ui(id)

mod_filters_server(input, output, session, rv, res_auth)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
