% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stats_values.R
\name{mod_stats_values_ui}
\alias{mod_stats_values_ui}
\alias{mod_stats_values_server}
\title{mod_stats_values_ui and mod_stats_values_server}
\usage{
mod_stats_values_ui(id)

mod_stats_values_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
